Add demonstration for Date & Time functionality provided in C++
In fact C++ inherits and makes use of what provided in C in <ctime> header,
it doesn't have its own date & time functionality as of C++11 as far as my knowledge.

Although there's some movement presented in C++20, but that's too far ahead, and
we can make use of <ctime> quite flexibly for now.