Add C & C++ version of BackdoorDefinitionWithNoDeclarationInLib
This demonstrate that definition of function (or variable) can be leaked if not careful
from implementation file.

C an C++ compiler, at least in this case gcc and g++ respectively handles and reports
such case differently.

C treats it as warning, but not for C++ which treats it as error. The former still allows us
to opt-in to treat it as error by include -Wimplicit-function-declaration as compilation flag.