Add NoWideString.cpp for validate output UTF-8 string on Windows
We can construct a std::string as UTF-8 string via u8 literal string
utility. But on Windows, we will be facing with issue of not being able
to print it out on console properly. Output will be messed looking like
it is still in different encoding.

We can instruct the executable binary of C++ program on Windows to
output string as UTF-8 by

- std::setlocale(LC_ALL, .UTF-8)
- SetConsoleOutputCP(CP_UTF8)
- compile the program with cl.exe with /utf-8 which means it will
  automatically add two options as of /execution-charset:utf-8 and
  /source-charset:utf-8

See https://learn.microsoft.com/en-us/cpp/build/reference/execution-charset-set-execution-character-set?view=msvc-170.